+++ udemyscraper/export/print.py
@@ -1,0 +1,9 @@
+def print_course(course):
+    print("===================== Fetched Course =====================", "\n")
+    print(course.title)
+    print(course.headline)
+    print(f"URL: {course.link}")
+    print(f"Instructed by {course.instructors}")
+    print(f"{course.rating} out of 5 ({course.no_of_ratings})")
+    print(f"Duration: {course.duration}")
+    print(f"{course.no_of_lectures} Lessons and {course.no_of_sections} Sections")
