+++ udemyscraper/metadata.py
@@ -1,5 +1,5 @@
 __package_name__ = 'udemyscraper'
-__version__ = "0.7.4"
+__version__ = "0.8.0"
 __author__ = 'Aditya Gupta'
 __author_email__ = 'divyashgupta2005@gmail.com'
 __description__ = """A Udemy Course Scraper built with beautiful soup, \
